重写科室，组合，明细接口

+++ backend/src/main/java/com/ljq/backend/dto/CombinationDTO.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class CombinationDTO {
-    private Integer id;               // 项目编号
+    private Long id;               // 项目编号
     private String name;              // 项目名称
     private Integer isActive;
     private String departmentName;    // 科室名称
-    private Integer departmentId;    // 科室编号
+    private Long departmentId;    // 科室编号
     private Integer gender;           // 性别
     private Double price;         // 单价
 }
