+++ backend/src/main/java/com/ljq/backend/entity/Combination.java
@@ -6,11 +6,11 @@ import java.math.BigDecimal;
 
 @Data
 public class Combination {
-    private Integer id;
+    private Long id;
     private String name;
     private Integer isActive;
     private Integer gender;
     private BigDecimal price;
-    private Integer departmentId;
+    private Long departmentId;
 
 }
