+++ backend/src/main/java/com/ljq/backend/entity/CombinationDetail.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CombinationDetail {
-    private Integer combinationId;
-    private Integer detailId;
+    private Long combinationId;
+    private Long detailId;
 }
