+++ backend/src/main/java/com/ljq/backend/entity/Detail.java
@@ -6,10 +6,10 @@ import java.math.BigDecimal;
 
 @Data
 public class Detail {
-    private Integer id;
+    private Long id;
     private String name;
     private boolean gender;
     private BigDecimal price;
-    private Integer departmentId;
+    private Long departmentId;
 
 }
