+++ backend/src/main/java/com/ljq/backend/mapper/PackageMapper.java
@@ -1,7 +1,7 @@
 package com.ljq.backend.mapper;
 
 import com.ljq.backend.dto.*;
-import com.ljq.backend.dto.page.PageDTO;
+import com.ljq.backend.dto.PageDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
@@ -11,12 +11,12 @@ public interface PackageMapper {
     List<PackageDTO> findAllPackages();
     List<PackageDTO> findPagePackages(PageDTO pageDTO);
     void updatePackage(PackageDTO packageDTO);
-    void deletePackage(Integer id);
+    void deletePackage(Long id);
     void addPackage(PackageDTO packageDTO);
-    PackageDTO findById(Integer packageId);
+    PackageDTO findById(Long packageId);
 
-    List<CombinationDTO> findSelectedItems(Integer packageId);
-    List<CombinationDTO> findAvailableDetails(Integer packageId);
+    List<CombinationDTO> findSelectedItems(Long packageId);
+    List<CombinationDTO> findAvailableDetails(Long packageId);
 
-    List<CombinationDTO> getCombinationsByPackageId(Integer packageId);
+    List<CombinationDTO> getCombinationsByPackageId(Long packageId);
 }
