+++ backend/src/main/java/com/ljq/backend/service/CombinationDetailService.java
@@ -3,5 +3,5 @@ package com.ljq.backend.service;
 import java.util.List;
 
 public interface CombinationDetailService {
-    void updateSelectedDetails(Integer combinationId, List<Integer> newDetailIds);
+    void updateSelectedDetails(Long combinationId, List<Long> newDetailIds);
 }
