+++ backend/src/main/java/com/ljq/backend/service/PackageCombinationService.java
@@ -3,5 +3,5 @@ package com.ljq.backend.service;
 import java.util.List;
 
 public interface PackageCombinationService {
-    void updateSelectedCombinations(Integer packageId, List<Integer> newCombinationIds);
+    void updateSelectedCombinations(Long packageId, List<Long> newCombinationIds);
 }
