+++ backend/src/main/java/com/ljq/backend/service/PackageService.java
@@ -3,7 +3,7 @@ package com.ljq.backend.service;
 import com.github.pagehelper.PageInfo;
 import com.ljq.backend.dto.CombinationDTO;
 import com.ljq.backend.dto.PackageDTO;
-import com.ljq.backend.dto.page.PageDTO;
+import com.ljq.backend.dto.PageDTO;
 
 import java.util.List;
 
@@ -11,10 +11,10 @@ public interface PackageService {
     List<PackageDTO> findAllPackages();
     PageInfo<PackageDTO> findPagePackages(PageDTO pageDTO);
     void updatePackage(PackageDTO packageDTO);
-    void deletePackage(Integer id);
+    void deletePackage(Long id);
     void addPackage(PackageDTO packageDTO);
-    List<CombinationDTO> findSelectedItems(Integer packageId);
-    List<CombinationDTO> findAvailableDetails(Integer packageId);
-    List<CombinationDTO> getCombinationsByPackageId(Integer packageId);
+    List<CombinationDTO> findSelectedItems(Long packageId);
+    List<CombinationDTO> findAvailableDetails(Long packageId);
+    List<CombinationDTO> getCombinationsByPackageId(Long packageId);
 
 }
