+++ backend/src/main/java/com/ljq/backend/service/impl/PackageCombinationServiceImpl.java
@@ -18,19 +18,19 @@ import java.util.stream.Collectors;
 public class PackageCombinationServiceImpl implements PackageCombinationService {
     private final PackageCombinationMapper packageCombinationMapper;
 
-    public void updateSelectedCombinations(Integer packageId, List<Integer> newCombinationIds) {
+    public void updateSelectedCombinations(Long packageId, List<Long> newCombinationIds) {
         // 1. 校验参数，使用局部变量存储 newCombinationIds
-        List<Integer> combinationIds = (newCombinationIds == null) ? Collections.emptyList() : newCombinationIds;
+        List<Long> combinationIds = (newCombinationIds == null) ? Collections.emptyList() : newCombinationIds;
         // 2. 查询旧的项目ID列表
-        List<Integer> oldCombinations = packageCombinationMapper.findCombinationIdsByPackageId(packageId);
+        List<Long> oldCombinations = packageCombinationMapper.findCombinationIdsByPackageId(packageId);
 
         // 3. 计算需要删除的项（旧列表存在，新列表不存在）
-        List<Integer> toDelete = oldCombinations.stream()
+        List<Long> toDelete = oldCombinations.stream()
                 .filter(id -> !combinationIds.contains(id)) // 使用局部变量 combinationIds
                 .collect(Collectors.toList());
 
         // 4. 计算需要新增的项（新列表存在，旧列表不存在）
-        List<Integer> toAdd = combinationIds.stream() // 使用局部变量 combinationIds
+        List<Long> toAdd = combinationIds.stream() // 使用局部变量 combinationIds
                 .filter(id -> !oldCombinations.contains(id))
                 .collect(Collectors.toList());
         // 5. 执行删除和新增
